Added image folder, portfolio picture and background, reworked css style system
Added a portfolio picture of me and a svg orange background decal. The decal is relatively positioned such that the absolutely positioned picture of me will be on top.

Added a html,body style to the css and removed the in-line one (see next comment). Most importantly, I added the overflow: hidden in-line style to the body element on line 16 of index.html. I used an in-line style because
1. If I do use the same css for multiple html files, I don't want all body elements to prevent overflow scroll
2. I did not want to make a class specifically for enabling and disabling scroll
3. For the other stylesheet that will be made in the future, I don't want my landing page to be scrollable there either, and the design made there will need to account for that. (Worst case I just remove the in-line)
Why am I being so effortful to write comments that only I will be looking at?

I was originally using px for nearly everything, but I switched to rem for padding, em for text, and vh/vw for positioning and scaling. Not only is this better conventionally, but this ensures the sizing of everything scales with the size of the browser. Also, I realized that I cannot be using a lot of in-line style in my html because the css will be modular, so at the cost of more classes/messier .css, we now have less in-line style in html. Considering doing multiple stylesheets: two for each html, each being of the different required styles.