Completed landing page, created introduction container, information container
The header, barring some positional changes, as been largely finished. I want to add a scrolling ticker with icons for my past projects to encourage users to click/explore, but that must happen after later website developments (time constraint)

Introduction header contains the Howdy, I'm Graham. The Graham text is absolutely positioned to bring it closer to the Howdy, I'm, and therefore the introduction div must use margins instead of just padding.

Information header uses a flex column display to keep the paragraphs and 'learn more' <a/>. Added a svg for the arrow, will probably have to be changed to white for adaptability to other css styles. A filter will then be needed to change color.