Completed AI page, added chaptgpt svg
Given how the svg logo for chatgpt is a one-off, its just kept normally as an icon in the svg folder, without a subfolder.

I am not entirely sure if the ai page was supposed to be ENTIRELY generated by AI or just have elements from AI, and I (for safety) assumed the later.