Made style change button persist across html pages and alerts on change (temporary)
The style button uses local storage to store its state and, on new page load, will update the state of any existing checkbox to align with the stored state.

Temporarily added an alert for style change as a second style doesn't exist yet.

Temporarily removed the animations from the style slider because it would move on new page load which looks bad.